+++ src/factories/pair/uniswap-pair.factory.ts
@@ -10,12 +10,12 @@ import { getTradePath } from '../../common/utils/trade-path';
 import { TradePath } from '../../enums/trade-path';
 import { UniswapVersion } from '../../enums/uniswap-version';
 import { UniswapContractContextV2 } from '../../uniswap-contract-context/uniswap-contract-context-v2';
-import { FeeAmount } from '../router/enums/fee-amount-v3';
 import { AllPossibleRoutes } from '../router/models/all-possible-routes';
 import { BestRouteQuotes } from '../router/models/best-route-quotes';
 import { RouteQuote } from '../router/models/route-quote';
 import { UniswapRouterFactory } from '../router/uniswap-router.factory';
 import { UniswapRouterContractFactoryV2 } from '../router/v2/uniswap-router-contract.factory.v2';
+import { FeeAmount } from '../router/v3/enums/fee-amount-v3';
 import { UniswapRouterContractFactoryV3 } from '../router/v3/uniswap-router-contract.factory.v3';
 import { AllowanceAndBalanceOf } from '../token/models/allowance-balance-of';
 import { Token } from '../token/models/token';
