+++ src/factories/uniswap-quoter/v3/uniswap-contract.quoter.v3.public.ts
@@ -1,5 +1,5 @@
-import { ChainId } from '../../enums/chain-id';
-import { EthersProvider } from '../../ethers-provider';
+import { ChainId } from '../../../enums/chain-id';
+import { EthersProvider } from '../../../ethers-provider';
 import { UniswapContractQuoterV3 } from './uniswap-contract.quoter.v3';
 
 export class UniswapContractQuoterV3Public extends UniswapContractQuoterV3 {
